// appfront/src/app/modules/employees/models/employeesDto/detailEmployeeDto.ts
export class DetailEmployeeDto {
    id: number | null = null;
    name: string | null = null;
    identification: string | null = null;
    urlImage: string | null = null;
    dateRegistry: Date | null = null;
    txtJobTitle: string | null = null;
    idJobTitle: string | null = null;
    base64: string | null = null;
    fileName: string | null = null;
}